+++ issues_statistics_test.go
@@ -34,8 +34,8 @@ func TestGetIssuesStatistics(t *testing.T) {
 	})
 
 	opt := &GetIssuesStatisticsOptions{
-		AssigneeID: Ptr(1),
-		AuthorID:   Ptr(1),
+		AssigneeID: Ptr(int64(1)),
+		AuthorID:   Ptr(int64(1)),
 	}
 
 	issue, _, err := client.IssuesStatistics.GetIssuesStatistics(opt)
@@ -67,8 +67,8 @@ func TestGetGroupIssuesStatistics(t *testing.T) {
 	})
 
 	opt := &GetGroupIssuesStatisticsOptions{
-		AssigneeID: Ptr(1),
-		AuthorID:   Ptr(1),
+		AssigneeID: Ptr(int64(1)),
+		AuthorID:   Ptr(int64(1)),
 	}
 
 	issue, _, err := client.IssuesStatistics.GetGroupIssuesStatistics(1, opt)
@@ -100,8 +100,8 @@ func TestGetProjectIssuesStatistics(t *testing.T) {
 	})
 
 	opt := &GetProjectIssuesStatisticsOptions{
-		AssigneeID: Ptr(1),
-		AuthorID:   Ptr(1),
+		AssigneeID: Ptr(int64(1)),
+		AuthorID:   Ptr(int64(1)),
 	}
 
 	issue, _, err := client.IssuesStatistics.GetProjectIssuesStatistics(1, opt)
